pyproject.toml
[project]
name = "ot-harjoitustyo"
version = "0.1.0"
description = "An application for managing your Pokémon cards"
authors = [
    {name = "Karstonner",email = "tomi.poutanen@helsinki.fi"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "invoke (>=2.2.0,<3.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "coverage (>=7.8.0,<8.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pylint (>=3.3.6,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
